updated tests file
