+++ tests/coftest.py
@@ -7,9 +7,3 @@ def app():
 @pytest.fixture
 def client(app):
     return app.test_client()
-
-def test_404(app, client):
-    res = client.get('/')
-    assert res.status_code == 200
-    expected = {'errorCode' : 404, 'message' : 'Route not found'}
-    assert expected == json.loads(res.get_data(as_text=True))
